+++ R/00-NAMESPACE.R
@@ -16,8 +16,8 @@
 #' @importFrom stats setNames
 #' @importFrom templates tmpl
 #' @importFrom tidyr separate unite
-#' @importFrom tools file_ext
+#' @importFrom urltools url_parse
 #' @importFrom utils capture.output download.file packageVersion
 #' @importFrom yaml  yaml.load_file
-#' @importFrom zip zipr
+#' @importFrom zip unzip zipr
 NULL